platsch: bugfix: use double precision in cairo_scale
cairo_scale parameters are the results of an integers division.
If the imported image size is greater than surface size, the
applied scale is 0 and nothing is drawn. If the image size
is smaller than the surface size, the image drawn doesn't fit
the surface properly. Scale parameters should be double precision
to allow for the image to fit properly the surface no matter the
scale between the image and the surface.

Upstream-Status: Pending

+++ cairo.c
@@ -115,7 +115,7 @@ static int png_import_backend_import_picture(cairo_t *cr, const char *filename)
 	surface_width = cairo_image_surface_get_width(surface);
 	surface_height = cairo_image_surface_get_height(surface);
 
-	cairo_scale(cr, surface_width/image_width, surface_height/image_height);
+	cairo_scale(cr, (double) surface_width/image_width, (double) surface_height/image_height);
 	cairo_set_source_surface(cr, image, 0, 0);
 	cairo_paint(cr);
